+++ comics/comics/eon.py
@@ -1,13 +1,16 @@
-from comics.aggregator.crawler import CrawlerBase, CrawlerImage
+from comics.aggregator.crawler import PondusNoCrawlerBase
 from comics.meta.base import MetaBase
 
 class Meta(MetaBase):
     name = 'EON'
     language = 'no'
-    url = 'http://www.nettavisen.no/tegneserie/striper/'
+    url = 'http://pondus.no/#CartoonGallery'
     start_date = '2008-11-19'
     rights = 'Lars Lauvik'
 
-class Crawler(CrawlerBase):
+class Crawler(PondusNoCrawlerBase):
+    history_capable_days = 14
+    schedule = 'Mo,Tu,We,Th,Fr,Sa,Su'
+
     def crawl(self, pub_date):
-        pass # Comic no longer published
+        return self.crawl_helper('EON', pub_date)