+++ comics/comics/darylcagle.py
@@ -13,6 +13,6 @@ class Crawler(CrawlerBase):
     time_zone = -5
 
     def crawl(self, pub_date):
-        url = 'http://www.cagle.com/working/%s/cagle00.gif' % \
-            pub_date.strftime('%y%m%d')
+        url = 'http://www.cagle.com/working/%s/cagle00.gif' % (
+            pub_date.strftime('%y%m%d'),)
         return CrawlerImage(url)