+++ docs/changes.rst
@@ -14,7 +14,9 @@ Unreleased
 - New crawler: ``atheistcartoons``
 - New crawler: ``petpeevy``
 - Add status page which for each comic shows when releases are fetched compared
-  to the comic's release schedule.
+  to the comic's release schedule. Contributed by Thomas Adamcik.
+- Add support for search in comic's title and text fields, using Haystack.
+  Contributed by Thomas Adamcik.
 
 **Bugfixes**