+++ docs/changes.rst
@@ -38,6 +38,7 @@ Unreleased
 - New: ``sheldon``
 - New: ``thegutters``
 - Update ``ctrlaltdelete`` to work after source site changes.
+- Update ``ctrlaltdeletesillies`` to work after source site changes.
 - Update ``dieselsweetiesweb`` to work after source site changes.
 - Update ``eon`` with new source site.
 - Update ``lunch`` with new source site.